+++ src/riak_cs_s3_response.erl
@@ -42,10 +42,11 @@
 
 -type xmlElement() :: #xmlElement{}.
 
--spec error_message(atom() | {'riak_connect_failed', term()}) -> string().
--spec error_code(atom() | {'riak_connect_failed', term()}) -> string().
--spec status_code(atom() | {'riak_connect_failed', term()}) -> pos_integer().
+-type error_reason() :: atom()
+                  | {'riak_connect_failed', term()}
+                  | {'malformed_policy_version', string()}.
 
+-spec error_message(error_reason()) -> string().
 error_message(invalid_access_key_id) ->
     "The AWS Access Key Id you provided does not exist in our records.";
 error_message(invalid_email_address) ->
@@ -73,6 +74,8 @@ error_message(admin_secret_undefined) -> "Please reduce your request rate.";
 error_message(bucket_owner_unavailable) -> "The user record for the bucket owner was unavailable. Try again later.";
 error_message(econnrefused) -> "Please reduce your request rate.";
 error_message(malformed_policy_json) -> "JSON parsing error";
+error_message({malformed_policy_version, Version}) ->
+    io_lib:format("Document is invalid: Invalid Version ~s", [Version]);
 error_message(malformed_policy_missing) -> "Policy is missing required element";
 error_message(malformed_policy_resource) -> "Policy has invalid resource";
 error_message(malformed_policy_principal) -> "Invalid principal in policy";
@@ -97,6 +100,7 @@ error_message(ErrorName) ->
     _ = lager:debug("Unknown error: ~p", [ErrorName]),
     "Please reduce your request rate.".
 
+-spec error_code(error_reason()) -> string().
 error_code(invalid_access_key_id) -> "InvalidAccessKeyId";
 error_code(access_denied) -> "AccessDenied";
 error_code(bucket_not_empty) -> "BucketNotEmpty";
@@ -116,6 +120,7 @@ error_code(bucket_owner_unavailable) -> "ServiceUnavailable";
 error_code(econnrefused) -> "ServiceUnavailable";
 error_code(malformed_policy_json) -> "MalformedPolicy";
 error_code(malformed_policy_missing) -> "MalformedPolicy";
+error_code({malformed_policy_version, _}) -> "MalformedPolicy";
 error_code(malformed_policy_resource) -> "MalformedPolicy";
 error_code(malformed_policy_principal) -> "MalformedPolicy";
 error_code(malformed_policy_action) -> "MalformedPolicy";
@@ -139,6 +144,7 @@ error_code(ErrorName) ->
 %% These should match:
 %% http://docs.aws.amazon.com/AmazonS3/latest/API/ErrorResponses.html
 
+-spec status_code(error_reason()) -> pos_integer().
 status_code(invalid_access_key_id) -> 403;
 status_code(invalid_email_address) -> 400;
 status_code(access_denied) ->  403;
@@ -161,6 +167,7 @@ status_code(multiple_bucket_owners) -> 503;
 status_code(econnrefused) -> 503;
 status_code(unsatisfied_constraint) -> 503;
 status_code(malformed_policy_json) -> 400;
+status_code({malformed_policy_version, _}) -> 400;
 status_code(malformed_policy_missing) -> 400;
 status_code(malformed_policy_resource) -> 400;
 status_code(malformed_policy_principal) -> 400;
@@ -205,6 +212,12 @@ api_error(Error, RD, Ctx) when is_atom(Error) ->
                    RD,
                    Ctx);
 api_error({riak_connect_failed, _}=Error, RD, Ctx) ->
+    error_response(status_code(Error),
+                   error_code(Error),
+                   error_message(Error),
+                   RD,
+                   Ctx);
+api_error({malformed_policy_version, _} = Error, RD, Ctx) ->
     error_response(status_code(Error),
                    error_code(Error),
                    error_message(Error),